+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Bump patch versions of Ruby 3.1, 3.2, 3.3 in the test matrix
 ### tmux
 - Add tmux 3.5a to test matrix
+### Features
+- Add active (-a) option to tmuxinator list
 
 ## 3.3.1
 ### Misc